backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pymupdf
python-docx
markdownify
